TabRecency: move to its own file
