+++ background_scripts/background.js
@@ -1,6 +1,7 @@
 import "../lib/utils.js";
 import "../lib/settings.js";
 import "../lib/url_utils.js";
+import "../background_scripts/tab_recency.js";
 import "../background_scripts/bg_utils.js";
 import "../background_scripts/commands.js";
 import "../background_scripts/exclusions.js";
