+++ tests/unit_tests/bg_utils_test.js
@@ -1,3 +1,4 @@
 import "./test_helper.js";
 import "../../lib/url_utils.js";
+import "../../background_scripts/tab_recency.js";
 import "../../background_scripts/bg_utils.js";
